Try to make day 7 faster
Starting time was 211 ms to solve part 2.

I fixed the calculation of sumCost to avoid an additional addition.
This reduced the time to 192ms.

I rewrote the cost summing to keep track of the accumulating costs as
it works through the list, and stop as soon as it exceeds a previous
cheaper cost.
This reduced the time to 73ms.